split interface & context from implementation with tests
